Update documentation on setting reno's `earliest-version` (#15243)
The important thing for `reno` is that it pulls in _all_ pre-releases in
the given minor cycle.  That doesn't always mean starting at `rc1`;
sometimes we make "beta" preview releases, and we still want to include
release notes that came in before those tags.  If we ever made an
"alpha" tag, we'd want to include that too.