Add methods to get active bits from QuantumCircuit
This commit adds two new methods to QuantumCircuit to get the indices of
active bits in the circuit. There was an existing rust space method to
get a tuple of sets of active bits, but that was not used for these new
public methods. This is because they use the bit objects instead of the
indices, and also we would want a deterministic iteration order and
using a PySet precludes this.

Additionally, the measure_active() method is updated to use this new
method internally because it was previously converting to a dag first to
compute the idle bits and then measuring the qubits outside that set.
Using this new method should be more efficient since it avoid the dag
creation.