Optimize the `OptimizeCliffordT` transpiler pass. (#14996)
* Porting the CliffordT optimization pass to Rust.

At this point the pass is still very naive and only cancels pairs of adjacent T-gates and pairs of
adjacent Tdg-gates.

A change in behavior: the pass raises an error if the circuit has non-(Clifford+T) gates.

* moving optimization code to a seperate functions

* Reimplementing Clifford+T optimization algorithm.

The optimization applies to sequences of 1-qubit Clifford+T/Tdg gates. We believe that for 1-qubit circuits
we get optimal T-counts.

* using StandardGate::X instead of X, as not to confuse with similarly-named Paulis

* Improved comments and adding Shelly as a co-author.

Co-authored-by: Shelly Garion <shelly@il.ibm.com>"

* removing unused imports

* missing i in the comment

* and removing 2

* This time properly adding Shelly as co-author

Co-authored-by: Shelly Garion <shelly@il.ibm.com>

* lint + avoid unnecessary collection

* changing const to static

* changing usize to u8

* adding release note

* adding a few additional sanity tests

* removing print

---------

Co-authored-by: Shelly Garion <shelly@il.ibm.com>