Move to rust native errors for rust apis in CircuitData
This commit migrates the CircuitData struct to have it's own new error
enum CircuitDataError. This error type encapsulates all the possible
error conditions returned from CircuitData methods and is used directly
in rust. This has two goals, first it enables us to return real error
messages to the C API, and secondly it decouples are rust code from
Python as part of #14240. To enable a build without pyo3 we need to
stop using PyErr in a pure rust context. The idea being that pymethods
will be conditionally compiled and not built, but the other methods that
aren't exposed to python will need to be returning rust error types. One
thing to note is there is at least one method still using a PyResult,
in those cases the methods will be conditionally compiled. Similarly
there is an ErrorFromPython variant in the new CircuitDataError, this
will also be conditionally compiled out, but is necessary because it's
used in places where we are calling python explicitly and need to handle
the error return from that.

Fixes #15027