Mark error enums as non-exhaustive and document their purpose
This commit adds the non_exhaustive attribute to all the error enums.
While we're using them there is the likelihood we'll add on to them with
new errors as methods are added. This also documents the enums are an
incremental step towards each function owning its error type as part of
the interface. These enums are just the first step towards that by
making it all explicitly defined. Eventually we'll want to split these
off into specific types that are specified per function/method instead
of being bundled together in a single enum.