Add option to filter inactive operations to active_qubits()
This commit adds a new option to the new methods,
filter_inactive_operations, which as the name implies will filter
inactive operations from being considered. This option is True by
default so that if you run .measure_active() is run it will return the
same qubits that measurements were added for.