Add descriptions of primitives container objects (#15292)
* Add descriptions of primitives container objects

This adds prose descriptions of the "container" and "handle" objects of
`qiskit.primitives`, which previously had stub documentation.  When
users are using the primitives and the returned objects interactively,
it can be very difficult to tell how to use each individual object,
especially because several of them are implemented using "magic
methods", which do not get suggested by code completion, and were not
alluded to in the per-class API documentation.

This PR doesn't include per-object examples - I didn't feel entirely
comfortable trying to fill that in myself - but the prose should
hopefully already be a significant improvement.

* Fix run-on inline code

* Correct source of `PrimitiveResult`

* Add discussion of `BitArray` constructors

Co-authored-by: Ian Hincks <2229105+ihincks@users.noreply.github.com>

---------

Co-authored-by: Ian Hincks <2229105+ihincks@users.noreply.github.com>