Split PyDAGCircuit from DAGCircuit
This commit splits out the Python only component of the DAGCircuit
struct in Rust. The DAGCircuit struct has two fields that explicitly
require Python, `metadata` and `duration`, and there are two more fields
that are only ever used in Python, `unit` and `name` despite both being
strings. To make the Python component distinct and separable this makes
a separate Python interface struct PyDAGCircuit which wraps the inner
pure rust DAGCircuit that contains the extra data for Python. This
struct also defines all the Python methods so the inner rust DAGCircuit
doesn't define any interface to Python anymore and it's all isolated to
the

This commit will prepare us towards implementing #14240 to make pyo3
and needing python symbols a compile time option. While it is possible
to conditionally compile all these aspects everywhere there are a few
advantages an explict stuct provides. The first is making all the places
we expect a Python interface very explicit. This makes it easy to reason
about special handling only needed for python and isolate it to separate
blocks. It also reduces the code complexity for when we do insert
conditional compilation features, because we only need to add the cfg
attributes to a few outer places instead of handling it internally in
every spot something from Python is used. Looking to the future the long
term goal this split enables is moving the Python aspects of the
DAGCircuit and the transpiler to the pyext crate. Having all the Python
interface components be separate structs and functions means when we
finish the data model migration all these interface points can just be
moved to the pyext crate since they'll be self contained. We can't do
that yet as control flow is still dependent on the Python interface.