Fix digit escaping in OpenQASM 3 identifier
The previous regex used `\w` as the match for an initial id character,
and `[\w\d]` for continuations.  Actually, `\w` matches any character
that `str.isalnum` returns `True` plus the underscore, which includes
all Unicode characters in the "digit" class, so is a superset of `\d`.
The match `\d` is only the Unicode group `[Nd]` (decimal digits), which
is not as complete as the entire `[N?]` set, so we can't even use
variants of `[^\W\d_]` or the like to get something usable; we'd have to
pull in the third-party `regex` package or the like to make an exact
match.

This instead causes the Qiskit exporter to allow only a subset of valid
OQ3 identifiers on export in lieu of _complete_ matching.  Additionally,
we now do not accidentally include digits in the allowed set of initial
characters for IDs (but do allow then in continuations).