Tune LitinskiTransformation pass
This commit optimizes the implementation of the LitinskiTransformation
pass. The improvements fall into a couple of categories, but most of the
runtime performance improvements will come from removing unecessary
allocations.

From a quick glance the Clifford struct looks like it could also use a
similar once over, mostly to get it to leverage Qiskit's types more
effectively if nothing else. But I kept the changes there minimal for
this commit and only modified the method used by The
LitinskiTransformation directly.