Skip intermediate PyList in PauliProductMeasurement::create_py_op() (#15310)
In the recently merged #15126 the PauliProductMeasurement operation
being added to Rust was unecessarily creating a pylist to go from
`Vec<bool>` -> `np.array[bool]` for the z and x parameters while we can
leverage rust-numpy to go straight from a Vec to a numpy array. This
was raised in review for the PR, but the fix just changed the mechanism
a Pylist was created by and didn't actually remove the python list
creation. `into_bound_py_any()` just is a short cut for using the
default IntoPyObject trait's conversion and casting it as a PyAny. In
the case of `Vec<T>` that means it's converting to a `PyList<T>`. Doing
this still results in 2 copies, one of which is done in Python. This
commit fixes the issue by using `to_pyarray()` to copy the Vec<bool>
directly into a numpy owned array. We have to copy still since we aren't
deleting the owning PauliProductMeasurement in Rust so we can't use
`into_pyarray()` which would avoid that copy. But this sill reduces the
numbers of copies needed to make the numpy arrays in Python.