Add `QkDag` instruction appliers and getter to C API
This adds `qk_dag_get_instruction` as a mirror to
`qk_circuit_get_instruction` (unifying the code in anticipation of
making it lower-allocation in the future), and the standard-instruction
application functions `qk_dag_apply_measure`, `qk_dag_apply_barrier` and
`qk_dag_apply_reset`.

We're still missing `qk_dag_apply_unitary`, but that's for a different
commit.