Add fixedbitset clifford for simd bitwise composition
This commit adds a custom clifford implementation for the litinski
transform that is simd optimized by using fixedbitset. This makes most
of the operations operating on the clifford use SIMD bitwise
instructions rather than iterating over arrays manually. This improves
the runtime performance of doing gate composition on the clifford. There
is likely a bit more overhead for iterating over the tableau but this
shouldn't be significant compared the the runtime gains constructing the
tableau.

For this commit the clifford class is duplicated because the other place
the existing struct in the quantum_info is used, clifford synthesis, is
based on the current tableau structure. This opted to copy the code
because I didn't want to change the synthesis code too. We should be
able to unify the implementations eventually, but it would require
adjusting anything directly accessing the clifford tableau which I
viewed as outside the scope of this PR.