Add C/Python interop for `Target` and `DAGCircuit`
This adds the ability to retrieve a `QkDag *` from a Python
`DAGCircuit` and a `QkTarget *`from a `Target`.  For `DAGCircuit`, which
is directly the user-facing pyclass, this commit also includes a way of
passing ownership from a `QkDag *` to a Python object.

Similar functionality could be added to `CircuitData` and
`SparseObservable`, but this is sufficient for immediate demonstrations
of transpiler passes.