No names for circuit library gate definitions (#15302)
* Make Qiskit less sharp

Make it less likely to accidentally end up with an opaque instruction that has the name of a standard gate, not naming the definition of a gate with the gate name. E.g. RXGate(..).definition.name shouldn't be "rx".

* Fix single controlled test

This test just happened to pass since the opaque gate we controlled was called "rz", triggering the "rz" path in add_control -- but in fact the gate represented a "p" gate. This is exactly the sort of facilitated foot-shooting this PR wants to avoid.

* add reno

* remove some additional name collisions

* Update releasenotes/notes/clib-definition-noname-75d0daf1355f4038.yaml

Co-authored-by: Alexander Ivrii <alexi@il.ibm.com>

---------

Co-authored-by: Alexander Ivrii <alexi@il.ibm.com>