Split out transpiler stages into separate rust functions (#15205)
* Split out transpiler stages into separate rust functions

For the rust function that implements the `qk_transpile()` C API it was
originally written as a single function where the full "staged pass
manager" was just a rust function and the pass manager logic was just
the logic inside that function. This worked fine and provided a full
pipeline as a single entrypoint which was the desire for qk_transpile.
However, with the plans to expose DAGCircuit to rust (see #15187) to
facilitate creating custom transpiler passes from C. To enable users to
pair custom passes with the default workflow it will be useful to
expose the preset stages to C as well.

This commit is the first step towards that, as it splits the stages
from the `qk_transpile()` into standalone rust functions. Each stage
gets it's own function which can be called independently. These new
functions have public visibility because the goal is that these will
be exposed to C. Following on from this PR separate PRs will be
opened for each stage that expose these new functions to the C API
and enable using them in custom transpilation workflows.

* Fix lint after rebase