Separate C version information into separate header (#15353)
* Separate C version information into separate header

This moves the large block of code into a proper header file so we can
use C tooling (like clang-format) on it.  More importantly, it begins
setting up a more complete structure for the `crates/cext/include`
directory so that `qiskit.h` itself can be a hand-written file, and
`cext` and `pyext` can separately write their own versions of the API
headers, which we'll need to have proper C extension modules in Python.

* Correct setup of ctests in coverage