Add `UnitaryGate` handling to `QkDag` C API
This adds methods to add and retrieve unitary matrices from `QkDag`.
We likely need something similar to retrieve the matrices from
`QkCircuit` as well at some point.

This commit uses copy-out semantics for the arrays to ensure that the
output is always row-major, but it's quite possible that we might want
an alternative method that fills in some sort of

```c
struct QkMatView {
    QkComplex64 *mat,
    size_t shape[2];
    size_t stride[2];
}
```

by just populating the pointer, and the metadata, so it's much cheaper.
We'd need to include the strides because Qiskit internally already can't
guarantee its matrices are contiguous, and the `OneQ`/`TwoQ` variants
are col-major while the `NDArray` variant is (usually) row-major.