Add substitute_node_with_dag to c api
This commit adds a new function qk_dag_substitute_node_with_dag to the C
api. This facilitates in place substitute from the new dag api. This is
fairly mechanical addition to the api as
`DAGCircuit::substitute_node_with_dag()` is already rust native. So
we only need to expose that inteface to C. The only difference between
the C and existing internal rust interface is that the qubit, clbit, and
var mapping arguments are not exposed to C. These options aren't
commonly used, and if the user needs to map the ordering they can do
that during replacement dag construction rather that at substitution
time.

Fixes #15190