Port control-flow operations to Rust. (#14568)
* Port control flow operations to Rust.

The change also moves ownership of basic blocks in Rust (the nested
circuits within control flow operations) to the containing circuit, i.e.
either CircuitData or DAGCircuit. As such, basic blocks now have the same
type as their container (no need to convert from CircuitData to a DAGCircuit
when you're accessing a control flow operation in a DAG context.

In Python, there is no change since we rebuild control flow operations
to avoid an API break.

Control flow instructions can be accessed by calling a
`try_view_control_flow` method of the circuit type, which provides the
following ergonomic view:

pub enum ControlFlowView<'a, T> {
    Box(Option<&'a Duration>, &'a T),
    BreakLoop,
    ContinueLoop,
    ForLoop {
        indexset: &'a [usize],
        loop_param: Option<&'a PyObject>,
        body: &'a T,
    },
    IfElse {
        condition: &'a Condition,
        true_body: &'a T,
        false_body: Option<&'a T>,
    },
    Switch {
        target: &'a Target,
        cases_specifier: Vec<(&'a Vec<CaseSpecifier>, &'a T)>,
    },
    While {
        condition: &'a Condition,
        body: &'a T,
    },
}

* Fix Rust unit tests.

* Clean up ObjectRegistery.

* Fix perf/memory regression.

I'd forgotten to add the logic to use None instead of an empty
smallvec in extract_params after moving to using an Option
for OperationFromPython::params.

This meant we were allocating and keeping around extra smallvecs
which was tanking circuit copy and extend (and ballooning our
peak memory usage).

* Fix params regression in one more place.

* Make create_py_op methods take ingredients by value.

This lets us get rid of an extra clone in parameter assignment,
introduced when unpack_py_op moved to the circuit level.

* Also use instruction::create_py_op in DAGCircuit::unpack_py_op.

* Don't panic in PackedInstruction param and block view methods.

We don't *need* to panic if someone tries to look at the params of a PackedInstruction
that happens to be a control flow instruction, or the blocks of a non-control flow
instruction. From the caller's perspective, PackedInstruction taking an "either or"
approach to params and blocks storage is more of an implementation detail.

The Parameters type itself still has 'unwrap_params' and 'unwrap_blocks' methods
which panic if you're wrong, but this makes more sense since it's an enum.

* Change params_view and blocks_view to avoid panic in Instruction, too.

* Implement get_params.

* Fix merge.

* Fix block deepcopy, add release note.

* Remove IntoPyObjectRef from Duration.

I'm not sure why I'd added this during development, but it
changes the behavior after PyO3 0.27 such that Durations
unwrap into the inner enum type, which is not what we
want.

* Fix Rust test.

* Remove CreatePythonOperation trait.

Instead, we just have the create_py_op function with owned args
to avoid hiding clones at the callsites.

* Make ControlFlowView::blocks return Vec.

* Use imports in operations.rs.

* Update Obj param comment.

* Add ControlFlowInstruction.

This is a struct that wraps ControlFlow so we can pull
qubits and clbits out from all the ControlFlow variants.

* Move ControlFlowView to operations.rs.

* Change bit type in Condition to bool.

* Use BigUInt in CaseSpecifier.

* Revert _ => to Obj(_).

* Parameters::unwrap_params and unwrap_blocks now takes.

* Add OperationFromPython::take_params and take_blocks.

* Favor Option<&...> over &Option<...> in params_eq.

* Address review comments in additional_wires.

* Document block_map.

* Remove TargetOperation::params.

* Use map function for qargs_mapping param.

* Address formatting comments.

* Don't use index in CircuitData API for instr methods.

* Add BlocksMode to make callsites clearer.

* Correct invalid-bounds message on variadic entries

* Remove unnecessary argument documentation

* Fix `BlocksMode` use in DAG test

* Rename `register_block` to `add_block`

* Run cargo fmt

* Rename `register_block` to `add_block`

* Use `&PackedInstruction` as lookup for `DAGCircuit::try_control_flow_view`

* Promote inner form of `unpack_py_op`

* Group no-op matches

* Clarify documentation comment

* Fix clippy complaints

---------

Co-authored-by: Jake Lishman <jake.lishman@ibm.com>