Correctly handle Blocks in `DAGCircuit::copy_empty_like`
For the Python-space function, we always want to "drop" blocks, because
there's no way to trigger block re-use from Python space, and we'd
always be appending them on afterwards. Rust, on the other hand, needs
to decide in each case how it wants to handle the blocks.  Using the
Python-space `copy_empty_like` was silently breaking things by using the
`BlocksMode::Drop` assumption.