Preserve Python `range` objects in for loops
The `ForLoopOp` always preserved Python `range` objects in the
structured form, and had special-purpose visualisation and OQ3 export
capabilities to ensure that these were handled in a structured way
end-to-end.  While the initial Rust port preserved the Python type
hints (`range` _is_ an `Iterable[int]`), it didn't preserve all the
intended behaviour.

This commit includes the setup to allow the for-loop collection to be
more dynamically typed as well.