Tune LitinskiTransformation pass (#15315)
* Tune LitinskiTransformation pass

This commit optimizes the implementation of the LitinskiTransformation
pass. The improvements fall into a couple of categories, but most of the
runtime performance improvements will come from removing unecessary
allocations.

From a quick glance the Clifford struct looks like it could also use a
similar once over, mostly to get it to leverage Qiskit's types more
effectively if nothing else. But I kept the changes there minimal for
this commit and only modified the method used by The
LitinskiTransformation directly.

* Remove unnecessary Vec for no clbits

* Use a more clear variable name for computing clifford count

* Add separate method for PPM inverse z

The earlier changes in this PR branch had modified the
`Clifford::get_inverse_z()` method to be optimized for building a sparse
observable. This however came at the cost of extra overhead for the
pauli product measurement path. The overhead came in two ways, it resulted
in an extra vec allocation to collect the bit terms, but also in an extra
iteration over all the terms to recompute the z and x components for the
pauli product measurement. This was mainly a bias left from the original
version of this PR which was written before PauliProductMeasurement was
introduced or used in the pass. This commit fixes this oversight by
adding a second method to Clifford for PPM usage of get_inverse_z that
returns the data in a similar format to what was there before seperate
pauli z and x Vecs.