Stop using a parallel sort in disjoint utils (#15410)
This commit switches the use of `par_sort_by_unstable()` to the serial
`sort_by_unstable()` implementation in the disjoint layout module. While
the parallel sort is faster and has some heuristics to limit the
multithreaded usage when the overhead isn't worth the tradeoff, the use
of multithreading from Python is potentially fraught when mixing with
multiprocessing in Python. For this reason all of our multithreaded
calls are supposed to guard against running in parallel based on env
vars that indicate the rust code is being executed in a multiprocessing
context. However the parallel sorts were missing this check and in this
case the overhead of checking we can run in multiple threads will likely
outweigh any speedup from a parallel sort. So this commit just opts to
drop the parallel sort and avoid the potential complications.