Fix control-flow builders with uncached Python control-flow
The `ElseContext` builder previously relied on referential identity of
the `IfElseOp`'s true body, which is no longer stable with the move of
control-flow operations to Rust space.  This sanity check is simpler and
easier to trick if you try, but it should still catch a proper set of
reasonable user errors.

It would be possible to add significantly more complex special handling
between Rust space and Python space to make the test less foolable, but
this commit is a reasonable trade-off for realistic user errors.