Add refcounting and free-list in control-flow blocks
This replaces the tracking of the control-flow blocks in `CircuitData`
and `DAGCircuit` with a custom stable-index list of blocks, imbibed with
a free list (implemented as an in-place stack), where each block counts
its number of references.  With the current control-flow setup, we do
not actually expect any valid circuit to increase the refcount of a
given block beyond 1, but since the stable-index and free-list form of
the object already required a custom `Entry` type to distinguish the
cases: present and referenced; present and not referenced; and absent;
the `refcount` field came for free anyway.

This updates `CircuitData` and `DAGCircuit` to keep these refcounts
up-to-date as the circuit is modified, and to immediately free blocks
whose refcount drops to zero (by default).

Automatic freeing of the blocks is necessary to prevent (effective)
memory leaks when calling methods like `DAGCircuit.substitute_node`,
which is a necessary call when a control-flow operation (like `BoxOp`)
is mutated from Python space to update it.