Fix timeline drawer for gates without unitary in target
The timeline drawer was storing the unitary of a gate from the target in
the metadata collected around drawing properties for some reason. This
field was never used by anything but was in the metadata dictionary used
for visualization. However, if there is a gate without a unitary in the
target (such as a parameterized gate like RGate(theta, phi) which has a
duration then this would error. Since we don't use the unitary for
anything there is no reason to collect. This commit fixes the issue by
removing the unitary from this internal metadata dictionary.