Add deprecation system to C API (#15369)
* Add deprecation system to C API

This adds infrastructure to apply deprecations to C API functions, which
causes them to raise warnings when user code is compiled against them
(with gcc/clang/MSVC, at least).  We also introduce a custom Doxygen
function that can be used to add deprecation information to the
docstring in a way that Sphinx will understand it.

To illustrate the use (and to follow, at least partially, the spirit of
our deprecation policies), this re-introduces the C API function
`qk_transpiler_pass_standalone_vf2_layout` function, which was removed
by the VF2-pruning PR chain.  We leave it with the Qiskit 2.2 API,
marking it as deprecated.

* Fix typos in documentation

Co-authored-by: Julien Gacon <gaconju@gmail.com>

* Comment on deprecation-warning location style

---------

Co-authored-by: Julien Gacon <gaconju@gmail.com>