Factor out block conversions dependent on foreign types
There is a lot of code in the `DAGCircuit` that depends on the foreign
types of the blocks within `CircuitData`, `OperationFromPython` and
`CircuitInstruction`.  This led to a lot of duplication of code, and
makes it costly to change the types of the stored blocks.  This is a
problem, because we already expect to change from `Py<PyAny>`
imminently, and then likely again to some other shareable type that will
permit copy-free exposure of the blocks to Python space.