Unify `ControlFlowView::try_from_instruction` (#15425)
`CircuitData` and `DAGCircuit` had two independent implementations of
this logic.  This originally stemmed from `CircuitData` using a
different internal storage mechanism for the control-flow blocks, but
since these were unified into the `ControlFlowBlocks` ref-counted
storage, it becomes easily possible to unify the control-flow viewing
logic.

This has a knock-on advantage of exposing a finer-grained borrow-check
model for the construction of the control-flow views, which is useful
for logic _within_ `CircuitData` (such as the `track_parameters` logic)
which wants to mutate part of the `CircuitData` state while still
producing these immutable control-flow borrows.