Unify symbol-extraction code in control-flow blocks parameter tracking (#15428)
* Unify symbol-extraction code in control-flow blocks parameter tracking

Previously, the `CircuitData::track_instruction_parameters` and
`untrack_instruction_parameters` logic was duplicated for extracting
symbols from blocks, which was particularly hard to update because it
involves coupled magic numbers.

This factors it out into a single "map" logic, which guarantees that the
two will always be in sync with each other.  This wasn't easily possible
earlier because of the borrow-check rules on how
`CircuitData::try_view_control_flow` worked, but with that factored out
to `ControlFlowView::try_from_instruction`, we can now have the
finer-grained borrow-check control that lets us do this.

A side effect of `ControlFlowView::try_from_instruction` is that we no
longer need to clone blocks in the way the algorithm is written, which
is more performant for construction.

* Document internal method for clarity