Move preferred location of `WrapAngles` default registry (#15101)
* Move preferred location of `WrapAngles` default registry

In Qiskit 2.2 we inadvertently exposed the default registry only from
the location
`qiskit.transpiler.passes.utils.wrap_angles.WRAP_ANGLES_REGISTRY`, which
traverses two private modules.  We had to document that location in
Qiskit 2.2 since it was already released when we spotted it, but had we
caught it before, we'd have changed the import path to something easier
and already public.

This PR attaches the default registry to the class; the `WrapAngles`
pass is called out as "special" from the `Target` as handling these
angle wrappings, so it seems appropriate for its default registry to be
attached there as well.

* Doc & reno review comments

---------

Co-authored-by: Julien Gacon <gaconju@gmail.com>