Allow control-flow block extraction to correct type
Previously, the `OperationFromPython` extraction helper automatically
chose `CircuitData` for the control-flow blocks.  Since this extraction
automatically required a clone anyway, the assumed type came at a cost.

This allows `DAGCircuit` to directly extract blocks to the correct type
without the intermediate step.  As a side effect, we also allow
extraction to fail with a type error if blocks are not expected, and add
a way to leave blocks unextracted.