Change `qargs_len` type to `uint32_t`
This avoids specifying the return value of `qargs_len` in the case of a
global operation, since we already commit to the only possible return
value of the pointer being null iff the properties are global.
Returning `uint32_t` is more consistent with our other use of
`qargs_len`, and we can still use `u32::MAX` as a length if we choose.