Change type of `CircuitData` blocks from `Py<PyAny>` to `CircuitData` (#15432)
* Change type of `CircuitData` blocks from `Py<PyAny>` to `CircuitData`

This removes one of the largest requirements for Python interaction in
the core data model.  However, it currently comes at a major cost of
cloning both on extraction from Python space and on producing view
objects back _to_ Python space.  We are choosing to make this trade-off
as a temporary measure, to prioritise separation of the core data model
from Python, given that transpiler performance in the presence of
control flow is already poor.

While `CircuitData` and `DAGCircuit` are both directly `pyclass`, we
either have to encode the Python interaction deep into the data
structures (deeply undesirable) to handle the blocks, or we have this
current problematic cloning.  The way around this will be to separate
off `PyCircuitData(Ptr<CircuitData>)` and similar for `DAGCircuit`,
where `Ptr<T>` is some smart-pointer type that may represent either
ownership of an underlying object or a (maybe fallible) reference to an
owned object.  The exact mechanism of this (perhaps something
`Arc`/`Weak`-based) is not yet decided, but will likely motivate a
further change of the `blocks` structure.

This commit is now a completely remade version of a previous PR
(gh-15123), but much of the set up work, investigation and planning of
what became this was originally done by Kevin Hartman.

Co-authored-by: Kevin Hartman <kevin@hart.mn>

* Commit crimes against default arguments in the name of easy determinism

* Update release note

* Make `assign_single_parameter` private

---------

Co-authored-by: Kevin Hartman <kevin@hart.mn>