Support simple `IfElseOp` conditionals in OpenQASM 2 export (#14556)
* Support simple `IfElseOp` conditionals in OpenQASM 2 export

With the removal of `Instruction.c_if` in gh-13506, we lost the ability
to export the sort of simple conditionals that OpenQASM 2 can represent.
This restores the ability, for the subset of `IfElseOp` objects that
OpenQASM 2 can represent.

It is possible to make the exporter slightly more permissive in some
cases, if desired/necessary.

* Specify supported CF in error message

---------

Co-authored-by: Kalyan Dasgupta <kalyan.dasgupta@gmail.com>
Co-authored-by: Julien Gacon <gaconju@gmail.com>