Clarify test assertion for `UnitarySynthesis`
In a rework of `UnitarySynthesis` I'm working on, I was (temporarily)
failing this test, but didn't have much to go on because the failure
message was just "False is not true".  This commit changes the assertion
to something more useful, so that the failing edges all appear in the
output.